% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Output}
\alias{Output}
\title{Output Object}
\usage{
Output(Output.outputMulti = NULL, id = NULL, outputLabel = NULL,
  outputMulti = NULL, outputValue = NULL, selfLink = NULL)
}
\arguments{
\item{Output.outputMulti}{The \link{Output.outputMulti} object or list of objects}

\item{id}{The unique name for the predictive model}

\item{outputLabel}{The most likely class [Categorical models only]}

\item{outputMulti}{A list of classes with their estimated probabilities [Categorical models only]}

\item{outputValue}{The estimated regression value [Regression models only]}

\item{selfLink}{A URL to re-request this resource}
}
\value{
Output object
}
\description{
Output Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Output functions: \code{\link{Output.outputMulti}}
}

