% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Training.modelInfo.confusionMatrix}
\alias{Training.modelInfo.confusionMatrix}
\title{Training.modelInfo.confusionMatrix Object}
\usage{
Training.modelInfo.confusionMatrix()
}
\value{
Training.modelInfo.confusionMatrix object
}
\description{
Training.modelInfo.confusionMatrix Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An output confusion matrix. This shows an estimate for how this model will do in predictions. This is first indexed by the true class label. For each true class label, this provides a pair {predicted_label, count}, where count is the estimated number of times the model will predict the predicted label given the true label. Will not output if more then 100 classes [Categorical models only].
}
\seealso{
Other Training functions: \code{\link{Training.modelInfo.confusionMatrixRowTotals}},
  \code{\link{Training.modelInfo}}, \code{\link{Training}},
  \code{\link{training.insert}}
}

