% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Training.modelInfo.confusionMatrixRowTotals}
\alias{Training.modelInfo.confusionMatrixRowTotals}
\title{Training.modelInfo.confusionMatrixRowTotals Object}
\usage{
Training.modelInfo.confusionMatrixRowTotals()
}
\value{
Training.modelInfo.confusionMatrixRowTotals object
}
\description{
Training.modelInfo.confusionMatrixRowTotals Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of the confusion matrix row totals
}
\seealso{
Other Training functions: \code{\link{Training.modelInfo.confusionMatrix}},
  \code{\link{Training.modelInfo}}, \code{\link{Training}},
  \code{\link{training.insert}}
}
