% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Training.modelInfo}
\alias{Training.modelInfo}
\title{Training.modelInfo Object}
\usage{
Training.modelInfo(Training.modelInfo.confusionMatrix = NULL,
  Training.modelInfo.confusionMatrixRowTotals = NULL,
  classWeightedAccuracy = NULL, classificationAccuracy = NULL,
  confusionMatrix = NULL, confusionMatrixRowTotals = NULL,
  meanSquaredError = NULL, modelType = NULL, numberClasses = NULL,
  numberInstances = NULL)
}
\arguments{
\item{Training.modelInfo.confusionMatrix}{The \link{Training.modelInfo.confusionMatrix} object or list of objects}

\item{Training.modelInfo.confusionMatrixRowTotals}{The \link{Training.modelInfo.confusionMatrixRowTotals} object or list of objects}

\item{classWeightedAccuracy}{Estimated accuracy of model taking utility weights into account [Categorical models only]}

\item{classificationAccuracy}{A number between 0}

\item{confusionMatrix}{An output confusion matrix}

\item{confusionMatrixRowTotals}{A list of the confusion matrix row totals}

\item{meanSquaredError}{An estimated mean squared error}

\item{modelType}{Type of predictive model (CLASSIFICATION or REGRESSION)}

\item{numberClasses}{Number of classes in the trained model [Categorical models only]}

\item{numberInstances}{Number of valid data instances used in the trained model}
}
\value{
Training.modelInfo object
}
\description{
Training.modelInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata.
}
\seealso{
Other Training functions: \code{\link{Training.modelInfo.confusionMatrixRowTotals}},
  \code{\link{Training.modelInfo.confusionMatrix}},
  \code{\link{Training}}, \code{\link{training.insert}}
}

