% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Update}
\alias{Update}
\title{Update Object}
\usage{
Update(classLabel = NULL, csvInstance = NULL)
}
\arguments{
\item{classLabel}{The true class label of this instance}

\item{csvInstance}{The input features for this instance}
}
\value{
Update object
}
\description{
Update Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Update functions: \code{\link{training.update}}
}
